refactor(ui): ♻️ refactor home ui and server-render events
- extract shared event-card component with alerts and time badges
- rename calltoaction to cta and refresh typography
- add loading skeleton and server-preloaded github stars
- server-render archive page and reserve scrollbar gutter
