+++ src/components/pages/home/cta.tsx
@@ -20,19 +20,21 @@ const CallToAction = () => {
         </div>
 
         <div className='relative z-10 flex flex-col text-left md:pr-[70px]'>
-          <div className='max-w-2xl'>
-            <h3 className='text-3xl font-semibold text-white md:text-4xl'>
-              Know a tech event? <br className='block' />{' '}
-              <span className='text-black'> Share it to help others find and join!</span>
+          <div className='max-w-2xl space-y-4'>
+            <h2 className='text-4xl font-bold tracking-tight text-white md:text-4xl'>
+              Know a tech event?
+            </h2>
+            <h3 className='text-2xl font-semibold text-black md:text-xl'>
+              Share it to help others find and join!
             </h3>
-            <p className='mt-3 text-lg font-medium text-white/80 drop-shadow-md'>
+            <p className='text-md leading-relaxed text-white/80'>
               Add your event to our list and be a part of the growing tech community.
             </p>
           </div>
 
-          <div className='mt-6 flex flex-wrap gap-4'>
+          <div className='mt-8 flex flex-wrap gap-6'>
             <button
-              className='rounded-lg bg-black px-6 py-3 text-base font-medium text-white transition-all duration-300 hover:scale-105 hover:bg-gray-900 focus:ring focus:ring-gray-400/40 active:scale-95'
+              className='text-md rounded-lg bg-black px-8 py-4 font-semibold text-white transition-all duration-300 hover:scale-105 hover:bg-gray-900 focus:ring focus:ring-gray-400/40 active:scale-95'
               onClick={() =>
                 window.open(
                   'https://github.com/FOSSUChennai/Communities/blob/main/.github/CONTRIBUTING.md'
@@ -42,7 +44,7 @@ const CallToAction = () => {
               Contribute
             </button>
             <button
-              className='rounded-lg border border-white/50 bg-transparent px-6 py-3 text-base font-medium text-white transition-all duration-300 hover:scale-105 hover:bg-white/20 focus:ring focus:ring-white/30 active:scale-95'
+              className='rounded-lg border-2 border-white/50 bg-transparent px-8 py-4 text-lg font-semibold text-white transition-all duration-300 hover:scale-105 hover:bg-white/20 focus:ring focus:ring-white/30 active:scale-95'
               onClick={() => router.push('/communities')}
             >
               Communities
